Takes communication code out of MPC320 class
